app_v2/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
PyYAML==6.0.1
aiofiles==23.2.0
gradio==5.44.1
requests==2.31.0
aiohttp==3.9.0
